Refactor some DTO class - Rename - package // domain.support -> dto
