+++ src/main/java/com/aoa/springwebservice/domain/Store.java
@@ -1,15 +1,13 @@
 package com.aoa.springwebservice.domain;
 
-import com.aoa.springwebservice.domain.support.MenuOutputDTO;
+import com.aoa.springwebservice.dto.MenuOutputDTO;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
-import org.hibernate.annotations.Cascade;
 import org.hibernate.annotations.Where;
 
 import javax.persistence.*;
-import javax.validation.constraints.Pattern;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
