+++ src/main/java/com/aoa/springwebservice/service/StoreService.java
@@ -3,9 +3,8 @@ package com.aoa.springwebservice.service;
 import com.aoa.springwebservice.domain.Store;
 import com.aoa.springwebservice.domain.StoreRepository;
 import com.aoa.springwebservice.domain.User;
-import com.aoa.springwebservice.dto.InputStoreDTO;
-import com.aoa.springwebservice.dto.OutputStoreDTO;
-import jdk.internal.util.xml.impl.Input;
+import com.aoa.springwebservice.dto.StoreInputDTO;
+import com.aoa.springwebservice.dto.StoreOutputDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -18,16 +17,16 @@ public class StoreService {
     @Autowired
     FileStorageService fileStorageService;
 
-    public Store createStore(InputStoreDTO storeDTO, User user){
+    public Store createStore(StoreInputDTO storeDTO, User user){
         return storeRepository.save(storeDTO.toDomain(saveStoreImg(storeDTO), user));
     }
 
-    public OutputStoreDTO convertToOutputDTO(Store store) {
-        return OutputStoreDTO.createOutputDTO(store);
+    public StoreOutputDTO convertToOutputDTO(Store store) {
+        return StoreOutputDTO.createOutputDTO(store);
     }
 
 
-    public String saveStoreImg(InputStoreDTO storeDTO) {
+    public String saveStoreImg(StoreInputDTO storeDTO) {
         return fileStorageService.storeFile(storeDTO.getImageFile());
     }
