+++ src/main/java/com/aoa/springwebservice/web/ApiStoreController.java
@@ -2,27 +2,15 @@ package com.aoa.springwebservice.web;
 
 import com.aoa.springwebservice.domain.Store;
 import com.aoa.springwebservice.domain.User;
-import com.aoa.springwebservice.dto.ExtendableDTO;
-import com.aoa.springwebservice.dto.InputStoreDTO;
-import com.aoa.springwebservice.dto.OutputStoreDTO;
-import com.aoa.springwebservice.response.ApiError;
-import com.aoa.springwebservice.response.ValidationErrorResponse;
+import com.aoa.springwebservice.dto.StoreInputDTO;
 import com.aoa.springwebservice.security.LoginUser;
 import com.aoa.springwebservice.service.StoreService;
-import lombok.extern.slf4j.Slf4j;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.BindException;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
-import javax.validation.Validation;
-import javax.validation.Validator;
-import javax.validation.ValidatorFactory;
-import java.net.URI;
 
 @RestController
 @RequestMapping("/api/stores")
@@ -33,10 +21,10 @@ public class ApiStoreController {
     StoreService storeService;
 
     @PostMapping("")
-    public String create(@Valid InputStoreDTO inputStoreDTO, @LoginUser User loginUser){
-        log.debug("inputDTO : {}", inputStoreDTO);
+    public String create(@Valid StoreInputDTO storeInputDTO, @LoginUser User loginUser){
+        log.debug("inputDTO : {}", storeInputDTO);
         //todo store 생성실패 상황 고려
-        Store store = storeService.createStore(inputStoreDTO, loginUser);
+        Store store = storeService.createStore(storeInputDTO, loginUser);
         return "/result/success";
     }
