+++ src/test/java/com/aoa/springwebservice/domain/MenuTest.java
@@ -1,6 +1,6 @@
 package com.aoa.springwebservice.domain;
 
-import com.aoa.springwebservice.domain.support.MenuDTO;
+import com.aoa.springwebservice.dto.MenuDTO;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.Test;
 import org.junit.runner.RunWith;
